Update nanomaterials flow (#1967)
* Only allow to notify product if nanom. is listed

When the user indicates that the nanomaterial is not listed in EC
regulation, the next step is an end page indicating them that they
cannot submit the product unless is listed in the annexes.

* Don't bypass steps unless nonstandard is only purpose

We want to follow the same steps when the purposes include or not a non
standard purpose as long as there is at least a standard purpose
selected.

* Only allow to notify product if nanom. conforms to restrictions

When the user indicates that the nanomaterial does not conform to the
restrictions set out in the annexes gets redirected to and end page
with a message explaining why cannot continue with the notification.

* Check non standard nanom. notification for multipurpose non standard

Once a multipurpose non standard nanomaterial has confirmed that is
listed in the regulations and conform to the restrictions for the
standard purposes it also has to confirm that has submitted a
notification for the nanomaterail non standard purpose.

* Update title for non standard nanom. notified step

- Replace UK with GB.
- Replace Brexit with date of the end of transition period.

* Update text on when can products with nanom. be placed on the market page.

- Replace dates and Brexit referrence with transition period end date.
- Replace UK with GB.

* Modify links on you can't notify product until nanom. is notified page

- Add a link ti the Notifications dashboard page.
- Open "how to notify nanomaterials" link in a new tab.

* Exclude non standard purposes from regulation annex details

Due to the flow changes we were reaching the pages that use this helper
with "other" (non standard) belonging to the nanomaterial purposes list.

As the non standard purpose doesn't have any regulation info, the
resultng text was "Annex X and ?". By these changes excluding non listed
Annex numbers for given purposes, the text turn into "Annex X" as
expected.

* Text changes

Upgrade text for "When products containing nanomaterial can be placed on
the market" page.

Replace hardcoded exit dates with constants.

* Confirm restrictions annexes link changes

Only show a link for the standard purposes.
Do not specify that they open in a new window.