Do not re-index Elastic search on prod boot
As our daba size increases, production start time of rails became
slower until the point of reaching the timeout for the health checks
and causing deployments to fail.

After tracing a Rails console start, was discovered that the long boot
time (a couple minutes at this point) was caused by ES indexing over
a db that keeps increasing in size.