Allow to store model attrs encrypted in db
Introduce a model concern that will allow to define one/multiple model
attributes as "encryptable".

Encryptable attributes will be stored encrypted in DB but recovered in
plain text from the model accessor.

This 2 way encryption (as opposite as 1 way hashing passwords) is needed
as we will need to be able to use the original value in some fields,
while we want to avoid store sensitive data in plain text on DB.

In the context for this is added, we will need to store a TOTP Secret
Key for users in DB, while we need the secret key value in order to
validate the TOTP Code provided by the user on the 2FA process.