Add&backfill user secondary authentication methods
As the users will be able to setup secondary authentication through a
sms dotp code, authentication app totp code or both, we need to register
wich method/s are setup for each user.

All the current users have been enforced to setup the sms dotp secondary
authentication, so the migration backfills the secondary authentication
methods to sms for them.