Enable App authenticator setup in security form
Instead of enforcing the mobile number to receive a security code by
sms, we offer the option to enable sms, app authenticator or both as
a secondary authentication method.

Depending on which method the user selects, they will have to:
- Introduce a valid mobile number.
- Setup the Authenticator App with the given QR Code or Secret key,
  then proceed to introduce the Authenticator App generated secret
  code.
- Do both.

Set 2FA cookie when registering with App Auth.

During the setup of the Authentication App on the account security page,
the user had to introduce the TOTP code in order to be able to submit
the account security form and finish their registration.

This should set the 2FA cookie.