Refactor: Reduce controller branching with OOP
The controlling is managing different logic depending if the user is
using SMS or APP authentication.

An obvious option would have been to split the controller into 2, one
for each 2FA method. But this would require to add redirection logic
on the secondary authentication concern and would require to use 2
different routes/paths for 2FA.

I personally like to keep "/two-factor" as a single gatekeep for 2FA,
but we may end splitting the controller anyway if we find this single
controller approach too complex.

To simplify the controller branching, I grouped the code that is shared
between both sms and app, and isolated what changes between both flows.

Ended creating a Handler abstraction that allows the controller flow to
be more linear and easy to follow while we can jump into the specific
handler to view/edit the particularities of each flow.

Kept the handler logic inside the controller as is only used there and
to avoid unnecessary indirection, as is easy to check the logic when is
kept together instead being stored in a different file.