Fix: Account Security secret key & QR changing
With the refactor of the TOTP code extracting it from the account
security code into time_otp model, the logic for keeping the
authenticator app secret key between form submissions attempts got
broken.

This was causing the QR code changing every time we submitted the
form. Obviously this is not acceptable as 99% of users won't notice
the code/key they scanned in their first attempt changed. So they
would keep attempting to submit codes generated with a wrong secret
key.

Fixed and covered it with specs.