Reduce RSpec output noise (#2052)
* Quiet the Sidekiq log output on spec runs

Sidekiq logging is throwing a lot of noise in the Rspec output.
This output is completely unnecessary during tests. If we need to check
the test output to the log... we should assert on it.

* Replace Devise deprecated bypass option on sign in

Triggering warning outputs on every spec run.
Replaced by "bypass_sign_in".

* Remove RSpec warning on possible false positives

Could have false positives as any other error being raised would
still cause the expectation to pass.
Better to just expect not to have any error to ensure the correct
run.