Fix analyzer exception when missing file
There are some cases where the blob reaches the ActiveStorage analyzer
not having an attachment associated.

There is code guarding in further steps for the case where the
Notification File hasn't been found. But the issue is being triggered
on the notification file search before this stage, so we need to guard
against it.

Seems this situation of "Blob without attachment" occurs  when the blob
has been removed from DB but is still on memory for ActiveStorage
analyzer purposes.
I guess this because I searched some of the ActiveStorage::Blob causing
the exceptions and none of them was found in DB.