April 1st 2021 changes (#2051)
* Remove pre-post brexit question for manual flow (#2037)

Default to "post brexit" behaviour enforcing the full manual flow.
Previously submitted notifications will still have the
"was_notified_before_eu_exit" flag in DB as "true", but new
notifications will rely on the default "false" value.

The question has been removed from the manual flow.

* Remove different behaviour for pre/post EU exit notifications (#2039)

* Enforce formulations & images on ZIP Notifications

Remove pre/post-brexit branching so both flows behave like what current
post-brexit flow does: Enforcing formulations and images to be present
in the notification.

* Add missing images as first Additional Info step

Both the image upload and the formulation upload triggered from
the additional information controller, redirect back to the additional
information controller once the action is finished.

This allows further missing information being requested from this
controller.

When nanomaterial missing information was being triggered before the
image upload, the submission of the component information redirected
the user to the pre-submission product overview without asking for
uploading the image. This was forcing the user to have to click again
on "Add missing information" for the notification in order to be
enquired for the image upload and be able to submit the notification.

Resolved by moving the image upload as the first thing to be triggered
from the Additional Information checks.