Bugfix: Default for cpnp_notification_date missing
When the cpnp_notification_date is missing in the parser, the importer
is raising an exception as tries to compare a "nil" value against a
date.

To resolve this, the comparison will only happen if
cpnp_notification_date is present, defaulting to a false value for
"Notification#was_notified_before_eu_exit" if not.