Fix flaky test on PH backlinks
The step asserting over the backlink for the last PH page step regularly
fails when the PH selection was "does not have a pH". Is not always but
happen frequently.

Changing the option to the default "between 3 and 10 pH" takes to the
expected page without failing in some of the spec runs.