Remove 2FA method selection from user session
When user has multiple methods configured for secondary authentication,
they need to choose which method to use when secondary authentication
is prompted.

The method selection was being set/read at the user session.

There were some situations where when abandoning the flow when prompted
for 2FA and then initiating a different flow that triggered a 2FA
required action, caused the previous method selection being used instead
of prompting the user again for selecting a method.

This can be avoided and simplified by remove the use of session to store
the method selection and just trust in the correct redirection happening
when the user selects a method.

So every time the user needs to fill 2FA it gets sent to the method
selection page and this form sends it to SMS or APP authentication page.