Enforce user having an 2FA request to resend SMS
In order to allow an user to resend its SMS code for 2FA, it needs to
have requested 2FA.

We can check the presence of an secondary authentication use id on the
user session, as this value gets set when 2FA is triggered.

Once that condition is met we don't need to worry about the redirection,
as when secondary authentication was previously propmted, it set the
redirection URL in the user session.