Ensure no previous user session when registering
There are some cases where the user partially signed in
(EG: following a "reset password link"), got queried for secondary
authentication and, instead of filling the 2FA form, went to the
homepage and attempted to Sign Up again.

When this happens, having some session info was causing issues,
like the user id for 2FA defaulting to the first user ID instead of
the newly registered one.

We need to be sure than when someone reaches the Sign Up form, its
previous season is wiped.