Cleans up ElasticS. indexes on review app deletion
This ensures we do not hit the shard limit on the Elasticsearch index
after a few hundred review apps have been deployed.

The task will fail silently if there are failures; this should not cause
much of a problem as long as it doesn't happen consistently.
But this ensures the review app will be deleted.