Allow RP members to edit Contact Person details
Member of a responsible person will be able to edit the details
(name, email address and phone number) of the Contact Person.

There was an "edit" action/view already defined for Contact Persons,
but wasn't called anywhere and the implementation was a single edit page
for all the CP details.

Our requirements is to show an independent link/page to edit each of the
details.