Remove status from redirections to /404
Redirecting with a non redirecting status causes the browser to
try to load the content into the body, showing a boilerplate text:
"You are being redirected".

There is no point to redirect to /404 pages with the not found status
when the redirection doesn't happen.
So better to redirect to 404 page but using the default redirection
status so the redirection occurs.