Fix Search bulk inviter issue with end of lines
The emails parsed from the CSV file contain end of line symbols,
like "user@example.org\n".

This causes issues when attempting to re-invite users as the user inviter
attempts to find an existent user by email, does not find it because the
end of line difference and then attempts to create a new user but fails
validation.

This happens silently and simply does not send the invitation.