Allow Responsible Person members to edit the RP contact person details (#2084)
* Adapt Summary Item spec matcher to accept new notation

Old summary item elements use tables.
New summary item helper use lists.

With this change, the matcher for tests can be used to assert over
summary items using either old or new notation.

* Allow RP members to edit Contact Person details

Member of a responsible person will be able to edit the details
(name, email address and phone number) of the Contact Person.

There was an "edit" action/view already defined for Contact Persons,
but wasn't called anywhere and the implementation was a single edit page
for all the CP details.

Our requirements are to show an independent link/page to edit each of the
details.

* Only display a confirmation message when changed

After submitting the form for editing a Contact Person field, we only
want to display the success confirmation message if the submitted value
is different from the original one.

* Restrict policy to edit RP Contact Person details

The "show?" policy allows users that are invited but didn't accept the
invitation to see the Responsible Person details.

In order to update the Contact Person details, the user should belong to
the Responsible Person. Invited users should not be able to edit
the Responsible Person information.

* Remove status from redirections to /404

Redirecting with a non-redirecting status causes the browser to
try to load the content into the body, showing a boilerplate text:
"You are being redirected".

There is no point to redirect to /404 pages with the not found status
when the redirection doesn't happen.
So better to redirect to 404 page but using the default redirection
status so the redirection occurs.