Refactor: Group all logic to New Email concern
Controller was knowing too much about of the internals of what should
happen when setting a new email for the user.

By grouping the logic into the New Email concern, the controller only
has the responsibility of:
- telling the user to set the new email pending confirmation.
- telling user class to confirm a new email given a token.

This also helps to trace the logic to set a new email now that is
grouped together under the same file.