Bugfix: Move mail delivery out of the DB transaction
BUG:
Users are requesting to change their email and receiving confirmation
links with invalid tokens.

These tokens seem to be the previous token that was changed with the
new email request. So, instead of receiving the new generated token,
users receive the token that was previously set and is no longer valid.

CAUSE OF THE BUG:
Change of token and mailer sending the new email confirmation link were
under the same DB transaction block.

EXPLANATION:
When code is wrapped into an Active Record transaction block, changes
are not written in DB until the full transaction block has been run
successfully.

This means that, when the mailer is called inside the transaction, even
when the "#save!" call has been called, the DB value has not been updated
as the transaction has not finished.

When this happens, the mailer retrieves the new email confirmation token
from DB and gets the previous token instead of the updated value, as the
update in DB happens when the transaction block ends.