Bugfix: New email confirmation link using outdated token (#2091)
* Bugfix: Move mail delivery out of the DB transaction

BUG:
Users are requesting to change their email and receiving confirmation
links with invalid tokens.

These tokens seem to be the previous token that was changed with the
new email request. So, instead of receiving the newly generated token,
users receive the token that was previously set and is no longer valid.

CAUSE OF THE BUG:
Change of token and mailer sending the new email confirmation link was
under the same DB transaction block.

EXPLANATION:
When code is wrapped into an Active Record transaction block, changes
are not written in DB until the full transaction block has been run
successfully.

This means that when the mailer is called inside the transaction, even
when the "#save!" call has been called, the DB value has not been updated
as the transaction has not finished.

When this happens, the mailer retrieves the new email confirmation token
from DB and gets the previous token instead of the updated value, as the
update in DB happens when the transaction block ends.

* Refactor: Group all logic to New Email concern

The controller was knowing too much about the internals of what should
happen when setting a new email for the user.

By grouping the logic into the New Email concern, the controller only
has the responsibility of:
- telling the user to set the new email pending confirmation.
- telling user class to confirm a new email given a token.

This also helps to trace the logic to set a new email now that is
grouped together under the same file.

* Refactor: Remove DB transactions

There is no need for them as we are just updating multiple fields in
the same table and there is no risk of data inconsistency.

By calling an update! we have an atomic update that will be reverted
if any of the fields do not pass validations.