Enforce Content Security Policy (CSP)
We have been running our CSP in "report only" mode.
With this mode, CSP violations get notified to Sentry but they are not
effectively blocked from being executed.

All our inline JS have been "approved" for CSP purposes through tagging
it with a nonce.
There are still a few violations but seem to be caused by extensions or
js code in the client side not injected by our service.

Due to that, seems we are ready to enforce the CSP, blocking the
violations.