Only allow to change mobile number when present
Only allow to change the mobile number when the user had set their
mobile number through the account security step.

If an user who only setup APP authentication goes to its profile
and tries to set their mobile number from the "change" link, gets
into a "broken" state when the user has a non verified mobile number
but cannot receive its SMS code as the user 2FA method is set to
what was selected on Account Security stage on registration (APP code).

Due to this, the user could not even navigate through the service, as
it would need to verify its mobile number but 2FA is not sending its
verification code.

Setting up changes on Account security is an incoming work, but for
the time being, we don't want to allow users who didn't set their phone
number on registration to be able to set it afterwards.