Move Mobile number change logic to model callback
The logic associated with the mobile number verification and adding
sms as 2FA method when a mobile number changes should not be commanded
from the controller.

This knowledge about what needs to happen when a mobile number changes
belongs to the user model.

In this approach the controller only needs to send a message to the user
to update its mobile number, and the user internally deals with what
needs to be done when setting/updating the mobile number.

As the associated logic is restricted to updating values in the same
user instance, and there are not external calls, dedided to go for a
model callback as it is ideal for this use case.