Move Sec. Auth. methods management to user callback
We completely remove from controllers/forms the responsibility of
setting/updating which secondary authentication methods are available
for the user.

The user model will be responsible for this logic.

Controllers/forms just need to update the sepecific user info like:
- User mobile phone.
- User totp secret key.

The implications from having the values above set regaring secondary
authentication methods available to the user are only known by the
user model.