Allow to navigate back from SMS to account security
When an user fills the account security page selecting Text message
as a secondary authentication method and submits the form, gets
redirected to the secondary authentication SMS page to confirm their
mobile phone number.

Currently, they cannot navigate back and they need to complete the
sms authentication.

Users are asking to be able to go back to the previous page to fill
the account security page only selecting the "App" authentication.

To achieve that, we're adding a backlink to a route that resets and
re-displays the user account security.