Remember previously selected security options
When the user follows a back link from the SMS code authentication page
after submitting the account security form, instead of displaying an
empty form where the user has to fill all the fields from scratch, we
will display the values previously selected by the user.

Password and Authenticator App code will be empty and needed to be
re-filled though.