Bugfix: Broken product image thumbnails in notification pages for search users (#2134)
* Enable ActiveStorage route for image thumbnails

Image thumbnails were broken for search users due to the active
storage route being disabled (re-routed to "/").

Enabled the route and overwrote the ActiveStorage representations
proxy controller to add an access policy to the images.

* Fix thumbnails URLs

URLs on the thumbnail images were pointing to Submit domain even when loaded from the Search service. This caused them not to be available for search users due to the newly introduced Pundit permissions checks.
Fixed by loading them with paths instead of using full URLs.