Fix invited submit users attempt to reset password
When a Submit Responsible Person invites a non registered user to their
team, and the invited user follows the invitation link, a new user gets
created in DB.

This new user needs to complete their registration with their name,
password and 2FA security.

Some users abandon the process without completing this form but seem
to be confused with their account status, as later they try to sign in
and follow the "reset password" process with their email.

Until now, the attempt to reset their password was failing as the user
was incomplete and missing the name, needed for the reset password email
personalisation.

We don't want to send them this email, we want to do as done with users
who registered themselves and didn't complete registration, sending them
the email to confirm their account instead.

In this case we will resend them the invitation email so they can
complete their registration.