Allow to edit images before complete notification
When, in the Submit service, a responsible person user registers a
notification and reaches the "preview before submit" ovewview page,
they will be able to follow a "Change" link for the notification label
image/s.

This link will send them to a notification image uploads edit page, from
where they will be able to remove the current image/s and add new one/s.

When this is submitted, the user will be sent back to the summary page.