Don't allow to change email to registered one
When an user attempts to change its email address to an email address
corresponding to an already registered account, the service is sending
the confirmation email but fails when attempting to confirm the change.

This is because user email addresses are unique and the update attempt
triggers an error.

What we want to do in those cases is to tell the user to wait for a
confirmation email but fail silently in the background and don't send
any email to the user.
This will avoid possible attackers discovering registered emails in the
service with the error messages.

We are also impeding registering email addresses for new users when the
email address is the new email (pending confirmation) for an existing
user.