Normalise interface for ActiveStorage access check
We have multiple classes using ActiveStorage attachments.
Some of these classes are associated to a notification, and the
notification belongs to a responsible person.

We also have classes that are directly associated to a responsible
person without the notification middleman class.

We are enforcing a common interface for the classes using ActiveStorage
access protection concern, so instead of querying each record inside
the concern, the concern will rely on the record responding to
"#responsible_person".