Add name to responsible person invitations
When inviting a person to the responsible person team, the invitation
form will require to provide the name of the invited person on addition
of their email address.

We are not introducing a name validation in the Pending Responsible
Person User model or in the database field due to having hundreds of
existing entries (existing invitations) that were created without a
name.