Fix Nanomaterial Notification attachments on seeds (#2226)
Seeded nanomaterial notifications where throwing an exception when
attempting to access their individual pages from the nano notifications
index page.
This was raised from the PDF file download link generation.

The cause is the lack of actual attachments on the Nanomaterial
Notifications seeds. As there is no attachment, the route generation
pointing to the attached file blows up.

Resolved attaching a PDF fixture for each of the seeded Nanomaterial
Notifications.