Reindex Elastic Search for each Review App deploy
Search Cosmetics on Review Apps is failing to load (500 error) due to
a missing Elastic Search index.
This forces developers to login into the review app console and manually
reindex ES.
We're automating this by adding a task triggering the reindex job and
calling that task on the review app deploy manifest.