Add format validation on Responsible Person names
As previously done on the user names, we wan't to restrict some
strings from being used in the Responsible Person names to avoid
possible spam issues.

We're creating a specific validator shared between the RP model and
the form used to introduce the RP name.

The model validation is only triggered when setting/changing the name
to avoid possible issues with existing data failing validations.