Extract parent class for name format validators
As the only difference between both Responsible Person and User format
validators is the regular expression rules used for validating the
format.

Not grouping them under a namespace/module as would need to use the
validations in this way:

validate :name, :'name_format/user' => true

instead of:

validate :name user_name_format: true

I personally find it silly so decided not to namespace it.