Remove unused file
I was wrong keeping this file when removing the CPNP logic from the
codebase during the Notification ZIP upload logic removal.

These particular methods and constants are not called from anywhere
outside this file. So seems it was exclusively related with the ZIP
upload logic and I got confused with the other notification_trigger_rules
file.