Remove unused trigger rules concern
The constants defined in this concern are also defined in the
notification_properties_helper.rb file.

These particular constants are not used anywhere, so there is no need
on keeping this concern around.

On top of that, because the definition of the constants inside the
"included { }" rails concern block, the constants definitions were
raising warnings on the spec runs.