Avoid double loading due to file naming
When the shared examples file ends with "_spec.b", it gets loaded twice
and triggers a warning.

Explanation on the rails_helper.rb file:

Files matching `spec/**/*_spec.rb` are
run as spec files by default. This means that files in spec/support that end
in _spec.rb will both be required and run as specs, causing the specs to be
run twice. It is recommended that you do not name files matching this glob to
end with _spec.rb.