Spec warning messages fixes (#2284)
* Remove unused file

I was wrong keeping this file when removing the CPNP logic from the
codebase during the Notification ZIP upload logic removal.

These particular methods and constants are not called from anywhere
outside this file. So seems it was exclusively related with the ZIP
upload logic and I got confused with the other notification_trigger_rules
file.

* Remove unused trigger rules concern

The constants defined in this concern are also defined in the
notification_properties_helper.rb file.

These particular constants are not used anywhere, so there is no need
on keeping this concern around.

On top of that, because the definition of the constants inside the
"included { }" rails concern block, the constants definitions were
raising warnings on the spec runs.

* Avoid warnings due to overwriting shared examples

Seems shared examples definitions, even when are defined inside
specific files/contexts, belong to the same global RSpec scope.

This causes that different shared contexts sharing the same name
clash with each other, and the last loaded context overrides the
previously loaded one.

This triggers a warning when running the specs.

Resolved it by using specific shared examples names for the context
where they are defined.

* Avoid double loading due to file naming

When the shared examples file ends with "_spec.b", it gets loaded twice
and triggers a warning.

Explanation on the rails_helper.rb file:

Files matching `spec/**/*_spec.rb` are
run as spec files by default. This means that files in spec/support that end
in _spec.rb will both be required and run as specs, causing the specs to be
run twice. It is recommended that you do not name files matching this glob to
end with _spec.rb.