Allow users to cancel invitations to their Responsible Person (#2318)
* Allow RP members to cancel team membership invites

Members of a Responsible Person will be able to cancel a team membership
invitation previously sent (but not accepted).

Once an invitation has been cancelled, it gets removed from DB and the
invitation link sent to the invited user won't be valid anymore.

* Move "Resend invitation" to invitations controller

Now that we have a specific controller for actions over Invitation
resources, the "resend invitation" action naturally falls under this
controller responsibility.

* Restrict invitation changes to Update policy

Only full team members of the Responsible Person will be able to operate
over the RP invitations.

* Move invitation creation to invitations controller

Having a dedicated controller for the invitations resource makes it
the right place for the new/create invitation actions.

* Remove unused route

Noticed this route does not map any controller action. So either it was
added by mistake or the action was removed from the controller.