Add a method to hard delete notifications
The usual ActiveRecord calls will still default to the soft deletion.
But now we will have a "#hard_delete!" call that will completely wipe
the instance (and dependent objects) from the Database, effectively
behaving as the original ActiveRecord "#destroy" method.
In fact, the "#hard_delete!" is just an alias for pre-overwritten
"ActiveRecord::Persistence#destroy" call.

If the notification receiving this call was soft deleted. It will be
removed together with their DeletedNotification associated object.