Remove model indirection for Deletion Service call
When a user tried to delete a notification, the controller was calling
a method in the notification model.
Then this instance method calls the NotificationDelete service object.

This method is exclusively called from the Deletion Controller, so we
can just move all the logic to the NotificationDelete class and call it
directly from the controller.