Fix: Redirect to 404 when notif. is not found
Search service was throwing 500 errors in some cases when users
attempted to visit a Notification page with an url containing the
wrong notification reference number.

This was caused by the controller not enforcing the notification to be
found prior to attempting to display it.

Fixed it by enforcing the retrieval of the notification. If the AR call
raises a not found error, this is mapped to the redirection to the 404
page.