Enable notification hard deletion (#2322)
* Refactor: Reword to soft delete for easier distinction

* Add a method to hard delete notifications

The usual ActiveRecord calls will still default to the soft deletion.
But now we will have a "#hard_delete!" call that will completely wipe
the instance (and dependent objects) from the Database, effectively
behaving like the original ActiveRecord "#destroy" method.
In fact, the "#hard_delete!" is just an alias for pre-overwritten
"ActiveRecord::Persistence#destroy" call.

If the notification receiving this call was soft deleted. It will be
removed together with their DeletedNotification associated object.

* Update specs to reflect soft/hard delete dichotomy

Two groups of specs:
- One for soft deletion
- One for hard deletion

Not changes in the spec logic itself, just description and what method
do we assert against (they're aliased so nothing changes).

* Remove model indirection for Deletion Service call

When a user tried to delete a notification, the controller was calling
a method in the notification model.
Then this instance method calls the NotificationDelete service object.

This method is exclusively called from the Deletion Controller, so we
can just move all the logic to the NotificationDelete class and call it
directly from the controller.