Upgrade from Ruby 2.7 to Ruby 3.0 (#2339)
* Upgrade CloudFoundry Ruby builpack to v1.8.49

* Upgrade to Ruby v3.0.3

* Upgrade rubocop to a recent version compatible with Ruby 3

* Rubocop autocorrections

* Rubocop set not auto-correctable issues as TODO

* Upgrade gem incompatible with Ruby 3 syntax

Active Hash gem latest version (3.1.5) is still incompatible with Ruby
3 new keyword arguments syntax.

Master is already fixed. So pointed the gem to a specific commit in
master in order to workaround the Ruby3 incompatibility.

Will upgrade it when a new version of the gem gets released.

* Upgrade Capybara to a version compatible with Ruby3

* Upgrade Strong Migrations gem

To resolve syntax compatibility issues with Ruby 3.

* Add Webrick into gemfile

Has been removed from Ruby 3 standard library/bundled gems.