Fix backlink for nanomaterial without nanoelements
The logic for the back link path generation was assuming that when a
nanomaterial is present in the component, there will be nanoelements
associated to the nanomaterial.

This is not the case for some imported notifications.
When these notifications are drafts, the user still can navigate through
their flow.

The service was throwing exception for those cases as
"last_nanoelement.standard?" was trying to call a method over a "nil"
object.

To resolve it, the branch condition asserts over the presence of a nano
element.