Guard against missing frame formulations IDs
When the wrong ID or SubID is given in the URL for the frame formulation
page, instead of assuming the formulation/subformulation are found and
trying to load the template, will reditect to 404 page, treating it
as an ActiveRecord not found situation.

Noticed this after an attempt of SQL attack against this controller,
passing SQL commands as ID/SUBID parameters.

The attack had no danger (specially as this controller does not query
the DB), but caused 500 errors in the service that brough attention over
the need to redirect to 404 instead.