Filter Pundit auth exceptions from Sentry (#2356)
* Filter Pundit auth exceptions from Sentry

These errors are meant to block users to access resources they shouldn't
be able to. The exceptions are not runtime exceptions that may be caused
by bugs or need to be resolved.

Given we have a very tight error quota in our Sentry plan and that Pundit
errors give us no value, we will benefit from not sending them to
Sentry.