Fix date helper called when date is not present
The construction of this line was causing the #display_full_month_date
helper method being called with nil values instead of only being called
when the notification completion date is present.

This is causing 500 errors in production when Search users attempt to
view notifications that lack completion date in DB (probably ZIP
imported).