Exclude deleted notifications from ESearch results
When a notification gets soft deleted, all their fields are set to nil.
If that notification was completed and indexed by ElasticSearch (ES),
ES will still return it as a search result for the Search service until
ES gets reindexed (once per day) and the notification, being in a state
different to completed, gets excluded from the new ES index.

During that period between the deletion and the next ES re-index, the
ES is associated to an invalid record (deleted state, emtpy fields).

This has recently caused displaying issues where the search results
index page triggers exceptions when trying to render empty fields,
breaking the search result page (500 error) for the user.

Instead of re-indexing every time that a notification gets deleted,
opening the door for a performance hit if a Responsible Person decides
to delete multiple notifications, we can accept that the index won't
be up to date for a while, and instead ensure that we only display the
search results that are associated to valid notifications.

We only add to the index completed notifications.
On the search results, we ensure they're still associated to completed
notifications before displaying them.
The ES results will get normalised on the next re-index job.