Fix search results limited to 10 by ElasticSearch
When the pagination was moved from the full search call to ES to the
record filtering, we inavertedly caused ElasticSearch results being
limited to 10 hits, its default.

We are fixing it by paginating the ES query results so we return the
expected 20 results per page, and then excluding non completed
notifications from the result.

We may find situations where a page shows less than 20 results, but
this is a minor downside rather than breaking the pagination,
limiting the results to 10 or attempting to display deleted records
that have been indexed by ElasticSearch.