Allow RP users to change Responsible Person type
Users from a Responsible Person will be able to edit the RP account
type, choosing between Individual or Business.

This change is accessible from the Responsible Person details page,
adding a link to edit the RP type.
This link sends the user to the RP edit page, also used for editing
the RP address.